--- Assets/Scripts/Races/Graphics/Mercs/Helldivers.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b5c8b4e3690131e448b4d5aa2da2a3c0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
