+++ Assets/Scripts/Races/Graphics/Mercs/RwuMercenaries.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a39647fd965744c4189cf6b3e95e0a50
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
