+++ Assets/Scripts/Races/Races.cs
@@ -49,7 +49,7 @@
     static internal Cockatrice Cockatrice = new Cockatrice();
     static internal Vargul Vargul = new Vargul();
     static internal Hamsters Hamsters = new Hamsters();
-    static internal Helldivers Helldivers = new Helldivers();
+    static internal RwuMercenaries RwuMercenaries = new RwuMercenaries();
 
     static internal Vagrants Vagrants = new Vagrants();
     static internal Serpents Serpents = new Serpents();
@@ -313,8 +313,8 @@
                 return Vargul;
             case Race.Hamsters:
                 return Hamsters;
-            case Race.Helldivers:
-                return Helldivers;
+            case Race.RwuMercenaries:
+                return RwuMercenaries;
             case Race.Dratopyr:
                 return Dratopyr;
             case Race.Erin:
