--- Assets/Sprites/Units/Helldivers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1a6b93d949491df47815370da5c7a957
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
