+++ Assets/Sprites/Units/RwuMercenaries.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6e50992de3d7ca944a10f58fbcf351db
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
